ui/pyproject.toml
[tool.poetry]
name = "genai-rtb"
version = "0.1.0"
description = ""
authors = ["Deepak Singh Rajput"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
streamlit = "^1.41.1"
pyarrow = "^19.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
